Move JSON files to data Folder
More tidying up.
