+++ src/assets/data/lang_korvax.json
