+++ src/views/ItemsView.vue
@@ -1,6 +1,6 @@
 <script>
 import { RouterLink } from 'vue-router'
-import itemsData from '../assets/items.json'
+import itemsData from '../assets/data/items.json'
 
 export default {
 	data() {
